Cleanup to use top level
