+++ ZipCmd/Program.cs
@@ -1,77 +1,54 @@
-﻿
-using Microsoft.Extensions.DependencyInjection;
+﻿using AbroadConcepts.CommandLine;
 using AbroadConcepts.IO;
+using Microsoft.Extensions.DependencyInjection;
+using ZipCmd;
 using ZipCmd.Models;
 using ZipCmd.Services;
-using AbroadConcepts.CommandLine;
 
-namespace ZipCmd;
-public static class Program
+var showHelp = true;
+var commandArguments = new CommandArguments(args);
+var mainArgument = new MainArgument();
+
+var zipCore = new ZipCore(commandArguments);
+var serviceCollection = ProgramHelper.SetupInitialServices(zipCore);
+var serviceProvider = serviceCollection.BuildServiceProvider();
+if (args.Length > 0 && commandArguments.Parse<IZipAction>(mainArgument, serviceProvider.GetServices<IZipAction>()))
 {
-    private static void Main(string[] args)
+    if (!string.IsNullOrEmpty(mainArgument.ZipFile))
     {
-        var showHelp = true;
-        var commandArguments = new CommandArguments(args);
-        var mainArgument = new MainArgument();
+        var filenames = mainArgument.ZipFile.EnsureExtension(".zip").GetFiles(false, true).ToList();
 
-        var zipCore = new ZipCore(commandArguments);
-        var serviceCollection = SetupInitialServices(zipCore);
-        var serviceProvider = serviceCollection.BuildServiceProvider();
-        if (args.Length > 0 && commandArguments.Parse<IZipAction>(mainArgument, serviceProvider.GetServices<IZipAction>()))
+        if (filenames.Count == 1)
         {
-            if (!string.IsNullOrEmpty(mainArgument.ZipFile))
+            using var stream = File.Open(filenames[0], FileMode.OpenOrCreate);
+            using var zipArchiver = new ZipArchiver(stream);
+            zipCore.Archiver = zipArchiver;
+            zipCore.Stream = stream;
+            var zipCommand = serviceProvider.GetService<IZipCommand>();
+            foreach (var option in commandArguments.Options)
             {
-                var filenames = mainArgument.ZipFile.EnsureExtension(".zip").GetFiles(false, true).ToList();
-
-                if (filenames.Count == 1)
-                {
-                    using var stream = File.Open(filenames[0], FileMode.OpenOrCreate);
-                    using var zipArchiver = new ZipArchiver(stream);
-                    zipCore.Archiver = zipArchiver;
-                    zipCore.Stream = stream;
-                    var zipCommand = serviceProvider.GetService<IZipCommand>();
-                    foreach (var option in commandArguments.Options)
-                    {
-                        showHelp = false;
-                        zipCommand?.Execute(option);
-                    }
-                }
-                else
-                {
-                    Console.WriteLine("ZipCmd - ZipFileName pattern does not reference a single filename.\n");
-                    foreach (var filename in filenames)
-                    {
-                        Console.WriteLine(filename);
-                    }
-                }
-            }
-            else
-            {
-                Console.WriteLine("No zip filename given.");
+                showHelp = false;
+                zipCommand?.Execute(option);
             }
         }
         else
         {
-            Console.WriteLine(commandArguments.Message);
-        }
-
-        if (showHelp)
-        {
-            var helper = serviceProvider.GetServices<IZipAction>().FirstOrDefault(s => s.GetType() == typeof(ZipHelp));
-            helper?.Execute();
+            showHelp = false;
+            ProgramHelper.ShowMultipleFilesError(filenames);
         }
     }
-
-    public static IServiceCollection SetupInitialServices(ZipCore core)
+    else
     {
-        return new ServiceCollection()
-            .AddTransient<IZipCommand, ZipCommand>()
-            .AddTransient<IZipAction, ZipAdd>()
-            .AddTransient<IZipAction, ZipExtract>()
-            .AddTransient<IZipAction, ZipList>()
-            .AddTransient<IZipAction, ZipRemove>()
-            .AddTransient<IZipAction, ZipHelp>()
-            .AddSingleton(core);
+        Console.WriteLine("No zip filename given.");
     }
+}
+else
+{
+    Console.WriteLine(commandArguments.Message);
+}
 
+if (showHelp)
+{
+    var helper = serviceProvider.GetServices<IZipAction>().FirstOrDefault(s => s.GetType() == typeof(ZipHelp));
+    helper?.Execute();
 }
