+++ ZipCmd/Services/ZipHelp.cs
@@ -14,10 +14,9 @@ public class ZipHelp(ZipCore core) : IZipAction
         Console.WriteLine("ZipCmd - Command line to operate on a zip file.");
         Console.WriteLine();
         Console.WriteLine("Currant passed arguments:");
-        Console.WriteLine(core.CommandArguments.ShowArguments());
+        Console.WriteLine($"\t{core.CommandArguments.ShowArguments()}");
         Console.WriteLine();
         Console.WriteLine("Command: ZipCmd.exe <ZipFilename> [-h] [-a <filePattern> <override> <compression> <entryLevel>]  [-r <filePattern> ]  [-e <filePattern> <directory> <overwrite]  [-l <filePattern> ]");
-        Console.WriteLine();
         Console.WriteLine();
         Console.WriteLine("Options:");
         Console.WriteLine("\t-a <filePattern> <override> <compression> <entryLevel> - Adds files to the zip file");
